+++ packages/ebayui-core/src/components/ebay-icon/icons/ebay-chase-18-colored-icon/symbol.ts
@@ -1,0 +1,4 @@
+export function symbol() {
+    // eslint-disable-next-line max-len,quotes
+    return '<symbol viewbox="0 0 30 18" id=icon-chase-18-colored><rect width=30 height=18 rx=2 fill=#005EB8></rect><path d="M13.16 2.531a.457.457 0 0 0-.458.457V6.19h8.457l-3.851-3.658h-4.149Zm8.309 4.628a.454.454 0 0 0-.457-.456h-3.2v8.46l3.655-3.855.002-4.149Zm-4.628 8.31a.457.457 0 0 0 .457-.458V11.81H8.84l3.852 3.658h4.15Zm-8.31-4.628c0 .253.205.459.458.459h3.2V2.84L8.532 6.691v4.15Z" fill=#fff></path></symbol>';
+}