+++ packages/ebayui-core/src/components/ebay-icon/icons/ebay-reply-chat-16-icon/symbol.ts
@@ -1,0 +1,4 @@
+export function symbol() {
+    // eslint-disable-next-line max-len,quotes
+    return '<symbol viewbox="0 0 16 16" id=icon-reply-chat-16><path d="M8 10a7 7 0 0 1-7-7 1 1 0 0 1 2 0 5 5 0 0 0 5 5h4.586l-2.293-2.293a1 1 0 1 1 1.414-1.414l4 4a1 1 0 0 1 0 1.414l-4 4a1 1 0 0 1-1.414-1.414L12.586 10H8Z"></path></symbol>';
+}