+++ packages/ebayui-core/src/components/ebay-icon/icons/ebay-stop-16-icon/symbol.ts
@@ -1,0 +1,4 @@
+export function symbol() {
+    // eslint-disable-next-line max-len,quotes
+    return '<symbol viewbox="0 0 16 16" id=icon-stop-16><rect x=1 y=1 width=14 height=14 rx=3></rect></symbol>';
+}