+++ packages/ebayui-core/src/components/ebay-icon/icons/ebay-stop-20-icon/symbol.ts
@@ -1,0 +1,4 @@
+export function symbol() {
+    // eslint-disable-next-line max-len,quotes
+    return '<symbol viewbox="0 0 20 20" id=icon-stop-20><rect x=2 y=2 width=16 height=16 rx=3></rect></symbol>';
+}